// Presentation/cosmeticshop.client/src/apiClient/models/cart-response-dto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * CosmeticShop.WebAPI
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: 1.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

import { CartItemResponseDto } from './cart-item-response-dto';
 /**
 * 
 *
 * @export
 * @interface CartResponseDto
 */
export interface CartResponseDto {

    /**
     * @type {string}
     * @memberof CartResponseDto
     */
    id: string;

    /**
     * @type {string}
     * @memberof CartResponseDto
     */
    customerId: string;

    /**
     * @type {number}
     * @memberof CartResponseDto
     */
    totalAmount: number;

    /**
     * @type {Array<CartItemResponseDto>}
     * @memberof CartResponseDto
     */
    cartItems: Array<CartItemResponseDto>;
}